Adding a "nextServer()" method to the Network model class.
This provides a means to incrementally and infinitely fetch Server instances. An Iterator instance was considered, but given the servers can be updated at any point, the instance would be liable to become inconsistent. The implementation allows for the list to be modified at any point with the method always returning a Server instance (assuming one is available..)