Adding a basic plugin framework to be integrated with the application.
This adds the skeletons for plugin construction. An abstract plugin class provides methods for hooking into events. Currently simple text matching is available for private messages, further methods for filtering events may be added at a later occasion. A PluginManager has been added, which deals with the loading of plugins, as they are expected to be provided by name from config. A FilteringCallback class provides a means to filter message events.