Application class now loads plugins via a manager instance.
Listeners are extracted from the plugins, and then registered via the standard route. The listener and subscriber registration methods have been refactored a little to simplify this process. The ApplicationConfig class now provides an instance of the PluginManager via the "core.plugin_manager" key, and sets the "plugins" and "plugins_path" keys with sensible defaults. The example "EchoBot" has been updated to reflect the fact we allow for multiple listeners per event name.